Robots work in parallel.
