+++ WarehouseSimulator.java
@@ -10,12 +10,14 @@ public class WarehouseSimulator {
             Warehouse w = new Warehouse(Warehouse.class.getResource("/warehouses/warehouse").getFile());
             Simulator s = new Simulator(w);
 
-            BlockingQueue<Factory.Package> packages = new ArrayBlockingQueue<>(w.getEntries().size());
-            Thread factoryThread = new Thread(new Factory(w.getEntries(), w.getDestinationSquares(), packages));
-            Thread directorThread = new Thread(new Director(w, s, packages));
+            BlockingQueue<Payload> payloads = new ArrayBlockingQueue<>(w.getEntries().size());
+            Thread factoryThread = new Thread(new Factory(w.getEntries(), w.getDestinationSquares(), payloads));
+            Thread transporterThread = new Thread(new DirectorTransporter(w, s, payloads));
+            Thread collectorThread = new Thread(new DirectorCollector(w, s, payloads));
 
             factoryThread.start();
-            directorThread.start();
+            collectorThread.start();
+            transporterThread.start();
         } catch (SimulatorException | GraphException ex) {
             ex.printStackTrace();
         }
