+++ simulator/Transporter.java
@@ -19,6 +19,7 @@ public class Transporter extends SimElement {
 
     private Image img;
     private boolean loaded;
+    private Payload payload;
 
 
     public Transporter(Coord startCoordinates) throws SimulatorException {
@@ -28,13 +29,23 @@ public class Transporter extends SimElement {
         setImages();
     }
 
-    public Image getImage() { return img; }
-
     public boolean isloaded() { return loaded; }
 
-    public void load() { loaded = true; }
+    public void load(Payload pkg) {
+        payload = pkg;
+        loaded = true; 
+    }
 
-    public void unload() { loaded = false; }
+    public void unload() { 
+        loaded = false;
+        payload = null;
+    }
+
+    public Payload getPayload() {
+        return payload;
+    }
+
+    public Image getImage() { return img; }
 
     public Coord getCoordinates() {
         return coord;
